pyproject.toml
[tool.poetry]
name = "firebasevezba"
version = "0.1.0"
description = ""
authors = ["Ivan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
firebase-admin = "^6.0.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
